feat: Refactor ConfigurableTool
