--- mirascope/tools/search/__init__.py
@@ -1,6 +1,0 @@
-from mirascope.tools.search.duckduckgo import (
-    DuckDuckGoSearch,
-    DuckDuckGoSearchConfig,
-)
-
-__all__ = ["DuckDuckGoSearch", "DuckDuckGoSearchConfig"]
