--- mirascope/tools/web/__init__.py
@@ -1,6 +1,0 @@
-from .parse_url_content import ParseURLConfig, ParseURLContent
-
-__all__ = [
-    "ParseURLContent",
-    "ParseURLConfig",
-]
